Refactor auth routes into auth_bp blueprint in the auth module
