+++ server/imageProcessor/auth/__init__.py
@@ -2,3 +2,4 @@
 Module for authorization
 """
 from .auth import token_required
+from .views import auth_bp
