+++ server/imageProcessor/views/index.py
@@ -2,7 +2,6 @@
 from flask import (session, redirect, url_for, request, abort, render_template,
                    send_from_directory, jsonify, make_response)
 from flask.blueprints import Blueprint
-from werkzeug.security import generate_password_hash, check_password_hash
 import requests
 import datetime
 import json
@@ -17,7 +16,7 @@ from imageProcessor.models import Item, ItemSchema, SavedItem, User, SearchLog, 
 from imageProcessor.auth import token_required
 from common import cred, constants, util
 
-api = Blueprint('views', __name__ )
+index_bp = Blueprint('index_bp', __name__)
 
 def get_words():
     words = set()
@@ -35,12 +34,12 @@ def get_companies():
             companies.add(word.strip())
     return companies
 
-@api.route('/')
+@index_bp.route('/')
 def hello_world():
     return make_response(jsonify({"message" : "hello world"}))
 
 
-@api.route('/items/', methods=["GET"])
+@index_bp.route('/items/', methods=["GET"])
 @token_required
 def get_saved_items(current_user):
     """
@@ -53,7 +52,7 @@ def get_saved_items(current_user):
     return make_response(jsonify(**context), 200)
 
 
-@api.route('/items/', methods=["PUT"])
+@index_bp.route('/items/', methods=["PUT"])
 @token_required
 def add_saved_item(current_user):
     """
@@ -107,7 +106,7 @@ def add_saved_item(current_user):
     return make_response(jsonify(**context), 201)
 
 
-@api.route('/items/', methods=["DELETE"])
+@index_bp.route('/items/', methods=["DELETE"])
 @token_required
 def delete_saved_item(current_user):
     """
@@ -130,7 +129,7 @@ def delete_saved_item(current_user):
     return make_response(jsonify({"error": "Item can't be deleted - item not found"}), 404)
 
 
-@api.route('/export_saved/', methods=["POST"])
+@index_bp.route('/export_saved/', methods=["POST"])
 @token_required
 def export_saved(current_user):
     """
@@ -142,7 +141,7 @@ def export_saved(current_user):
     return make_response(jsonify(**context), 201)
 
 
-@api.route('/history/', methods=["GET"])
+@index_bp.route('/history/', methods=["GET"])
 @token_required
 def get_history(current_user):
     context = {}
@@ -167,8 +166,7 @@ def construct_search_terms(search_terms, annotations, count=float('inf')):
                 search_terms.insert(0, term['description'].lower())
 
 
-
-@api.route('/search/', methods=["POST"])
+@index_bp.route('/search/', methods=["POST"])
 @token_required
 def search(current_user):
     """
@@ -285,87 +283,4 @@ def search(current_user):
     schema = ItemSchema(many=True)
     items_dict = schema.dump(items)
     context["items"] = items_dict.data
-    return make_response(jsonify(**context), 201)
-
-@api.route('/login/', methods=["POST"])
-def login():
-    """
-    Logs a user in, and returns a json web token for future authenticated requests
-    """
-    username = request.get_json(silent=True).get("username")
-    email = request.get_json(silent=True).get("email")
-    password = request.get_json(silent=True).get("password")
-    if not (username or email) or not password:
-        abort(400) # TODO: Handle Error - invalid request
-
-    user = None
-    if username:
-        user = User.query.filter_by(username=username).first()
-    elif email:
-        user = User.query.filter_by(email=email).first()
-
-    if not user:
-        # TODO: Handle Error - invalid username/password
-        abort(400)
-
-    if not check_password_hash(user.password_hash, password):
-        # TODO: Handle Error - invalid username/password
-        abort(400)
-
-    # Return json web token
-    token = user.get_token()
-    context = {}
-    context['username'] = user.username
-    context['token'] = token.decode('UTF-8')
-    return make_response(jsonify(**context), 200)
-
-
-@api.route('/logout/', methods=["POST"])
-@token_required
-def logout(current_user):
-    # TODO: Invalidate token somehow
-    return make_response(jsonify({"message": "Logged out"}), 200)
-
-
-@api.route('/users/', methods=["POST"])
-def create_user():
-    """
-    Creates a user and returns a jwt for future authenticated requests
-    """
-    # TODO: Verify fields
-    if not request.json \
-        or not 'username' in request.json \
-        or not 'email' in request.json \
-        or not 'password' in request.json \
-        or not 'firstname' in request.json \
-        or not 'lastname' in request.json:
-        return jsonify({"error": "invalid format"}), 400
-
-    # Check user doesn't already exist
-    user = User.query.filter_by(username=request.json['username']).first()
-    if user:
-        return jsonify({"error": "user already exists"}), 400
-    user = User.query.filter_by(email=request.json['email']).first()
-    if user:
-        return jsonify({"error": "user already exists"}), 400
-
-    # Insert user into db
-    password_hash = generate_password_hash(request.json['password'])
-    public_id = util.get_uuid()
-
-    user = User(
-        public_id=public_id,
-        username=request.json['username'],
-        email=request.json['email'],
-        password_hash=password_hash,
-        firstname=request.json['firstname'],
-        lastname=request.json['lastname'])
-    db.session.add(user)
-    db.session.commit()
-
-    # Return json web token
-    token = user.get_token()
-    context = {}
-    context['username'] = request.json['username']
-    context['token'] = token.decode('UTF-8')
     return make_response(jsonify(**context), 201)
